Updated review comments
