--- renderer/components/dashboard-projects/dashboard-filter/technologies-list.ts
@@ -1,14 +1,0 @@
-export const technologies = [
-  {
-    name: 'Node',
-    value: 'node',
-  },
-  {
-    name: 'Angular',
-    value: 'angular',
-  },
-  {
-    name: 'Java',
-    value: 'java',
-  },
-];
