+++ renderer/models/project-filter/project-filter.ts
@@ -1,0 +1,5 @@
+import { ProjectDetails } from '../../modules/projects/redux/stepper/data.model';
+
+export interface ProjectFilter {
+  filter(compareTo: string): ProjectDetails[];
+}
