+++ renderer/models/project-filter/project-sort.ts
@@ -1,0 +1,5 @@
+import { ProjectDetails } from '../../modules/projects/redux/stepper/data.model';
+
+export interface ProjectSort {
+  sort(): ProjectDetails[];
+}
